# code/backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Impax Backend Swagger",
    "version": "0.1.0",
    "description": "Welcome to ImpaX, where innovation meets the field of sports, prioritizing the well-being of athletes. Our mission is to revolutionize athlete safety through cutting-edge impact monitoring technology.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/cepdnaclk/e19-3yp-impact-tracker/blob/main/LICENSE"
    },
    "contact": {
      "name": "Imapax",
      "url": "https://github.com/cepdnaclk/e19-3yp-impact-tracker",
      "email": ""
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/auth": {
      "get": {
        "tags": ["AuthEndpoints"],
        "summary": "Authenticate a user - Refresh Token required",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/LoginResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/team": {
      "post": {
        "tags": ["TeamEndpoints"],
        "summary": "Create a team - Open",
        "requestBody": {
          "$ref": "#/components/requestBodies/TeamCreateRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/Team"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/team/exists/teamId/{id}": {
      "get": {
        "tags": ["TeamEndpoints"],
        "summary": "Check if a team ID exists - Open",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTeamId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/TeamIdExistsResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },
    "/team/exists": {
      "get": {
        "tags": ["TeamEndpoints"],
        "summary": "Check if both Team ID and email exists - Open",
        "parameters": [
          {
            "name": "teamId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTeamId"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example@example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/TeamIdEmailExistsResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/manager/exists/email/{email}": {
      "get": {
        "tags": ["ManagerEndpoints"],
        "summary": "Check if a manager email exists - Open",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example@example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/ManagerExistsResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/manager": {
      "post": {
        "tags": ["ManagerEndpoints"],
        "summary": "Create a manager - Open",
        "requestBody": {
          "$ref": "#/components/requestBodies/ManagerCreateRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/ManagerResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/manager/add": {
      "post": {
        "tags": ["ManagerEndpoints"],
        "summary": "Add a new manager to a team - Access Token required",
        "requestBody": {
          "$ref": "#/components/requestBodies/ManagerTeamRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/ManagerTeamRequest"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/login/logout": {
      "post": {
        "tags": ["LoginEndpoints"],
        "summary": "Logout - Access Token required",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "message": "Logout successful"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/login/manager": {
      "post": {
        "tags": ["LoginEndpoints"],
        "summary": "Login as a manager - Open",
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/LoginResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/player/add": {
      "post": {
        "tags": ["PlayerEndpoints"],
        "summary": "Add a new player to a team for manager - Access Token required",
        "requestBody": {
          "$ref": "#/components/requestBodies/PlayerTeamRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/PlayerTeamRequest"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    },

    "/login/player": {
      "post": {
        "tags": ["LoginEndpoints"],
        "summary": "Login as a player - Open",
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/LoginResponse"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Bad Request"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        },
        "example": {
          "password": "securePassword",
          "userName": "example@example.com"
        }
      },

      "ManagerTeamRequest": {
        "type": "object",
        "properties": {
          "managerEmail": {
            "type": "string"
          },
          "teamId": {
            "type": "string"
          }
        },
        "example": {
          "managerEmail": "example@example.com",
          "teamId": "teamId"
        }
      },

      "PlayerTeamRequest": {
        "type": "object",
        "properties": {
          "playerEmail": {
            "type": "string"
          },
          "teamId": {
            "type": "string"
          }
        },
        "example": {
          "playerEmail": "example@example.com",
          "teamId": "teamId"
        }
      },

      "LoginResponse": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        },
        "example": {
          "refreshToken": "abcd",
          "accessToken": "abcd"
        }
      },

      "ManagerExistsResponse": {
        "type": "object",
        "properties": {
          "managerExists": {
            "type": "boolean"
          }
        },
        "example": {
          "managerExists": true
        }
      },

      "ManagerCreateRequest": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "teamId": "exampleTeamId",
          "firstName": "John",
          "lastName": "Doe",
          "email": "john.doe@example.com",
          "password": "securePassword"
        }
      },

      "ManagerResponse": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "example": {
          "teamId": "exampleTeamId",
          "firstName": "John",
          "lastName": "Doe",
          "email": "john.doe@example.com"
        }
      },

      "TeamIdExistsResponse": {
        "type": "object",
        "properties": {
          "teamExists": {
            "type": "boolean"
          }
        },
        "example": {
          "teamExists": true
        }
      },
      "TeamManagerInterface": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "example": {
          "teamId": "exampleTeamId",
          "email": "example@example.com"
        }
      },
      "TeamIdEmailExistsResponse": {
        "type": "object",
        "properties": {
          "teamExists": {
            "type": "boolean"
          },
          "managerExists": {
            "type": "boolean"
          }
        },
        "example": {
          "teamExists": true,
          "managerExists": false
        }
      },

      "Team": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "teamName": {
            "type": "string"
          },
          "teamManager": {
            "type": "string"
          }
        },
        "example": {
          "teamId": "exampleTeamId",
          "teamName": "Example Team",
          "teamManager": "Team Manager email"
        }
      }
    },

    "requestBodies": {
      "TeamCreateRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Team"
            }
          }
        }
      },

      "ManagerTeamRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ManagerTeamRequest"
            }
          }
        }
      },

      "PlayerTeamRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PlayerTeamRequest"
            }
          }
        }
      },

      "LoginRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginRequest"
            }
          }
        }
      },

      "LoginResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginResponse"
            }
          }
        }
      },

      "ManagerCreateRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ManagerCreateRequest"
            }
          }
        }
      },

      "TeamIdEmailCheckRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TeamManagerInterface"
            }
          }
        }
      }
    }
  }
}